refactor: move advancement package out of blocks??
