+++ src/main/java/com/aetherteam/aether_genesis/Genesis.java
@@ -2,7 +2,7 @@ package com.aetherteam.aether_genesis;
 
 import com.aetherteam.aether.AetherConfig;
 import com.aetherteam.aether_genesis.block.GenesisBlocks;
-import com.aetherteam.aether_genesis.block.advancement.GenesisAdvancementTriggers;
+import com.aetherteam.aether_genesis.advancement.GenesisAdvancementTriggers;
 import com.aetherteam.aether_genesis.blockentity.GenesisBlockEntityTypes;
 import com.aetherteam.aether_genesis.blockentity.GenesisMenuTypes;
 import com.aetherteam.aether_genesis.client.GenesisSoundEvents;
