+++ src/main/java/com/aetherteam/aether_genesis/advancement/GenesisAdvancementTriggers.java
@@ -1,4 +1,4 @@
-package com.aetherteam.aether_genesis.block.advancement;
+package com.aetherteam.aether_genesis.advancement;
 
 import net.minecraft.advancements.CriteriaTriggers;
