+++ src/main/java/com/aetherteam/aether_genesis/item/materials/ContinuumOrbItem.java
@@ -1,7 +1,7 @@
 package com.aetherteam.aether_genesis.item.materials;
 
 import com.aetherteam.aether.item.miscellaneous.ConsumableItem;
-import com.aetherteam.aether_genesis.block.advancement.ContinuumOrbLootTrigger;
+import com.aetherteam.aether_genesis.advancement.ContinuumOrbLootTrigger;
 import com.aetherteam.aether_genesis.loot.GenesisLoot;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
